Exercise 4.21 complete

+++ Part4/bloglist/controllers/blogs.js
@@ -41,6 +41,17 @@ blogsRouter.post("/", async (request, response, next) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.VITE_TOKEN_KEY);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+      return response.status(404).json({ error: "blog not found" });
+    }
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(403).json({ error: "unauthorized access" });
+    }
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
   } catch (exception) {
